Add new migrations

+++ db/migrate/20170706210056_create_plans.rb
@@ -2,7 +2,7 @@ class CreatePlans < ActiveRecord::Migration[5.1]
   def change
     create_table :plans do |t|
       t.references :monthly_plan
-      t.integer :subcategory_id
+      t.bigint :subcategory_id
       t.decimal :amount
 
       t.timestamps
