+++ db/migrate/20170706212635_create_default_subcategories.rb
@@ -2,7 +2,7 @@ class CreateDefaultSubcategories < ActiveRecord::Migration[5.1]
   def change
     create_table :default_subcategories do |t|
       t.references :user
-      t.integer :subcategory_id
+      t.bigint :subcategory_id
 
       t.timestamps
     end
